Initial commit
This is a dummy program that highlights an issue with reading namelists
from a string. The NAG compiler doesn't like reading a string that
contains the carriage return (achar(10)) character, and PGI can't find
the namelist in the string. Switching to an array of strings, each
containing a single namelist and no carriage returns, highlights an
issue with gfortran (at least in 4.8.3) where it does not return a
status code if the namelist is not found in the string.