Write namelist at end of program
We are running into an issue when we try to write a namelist to a string and
then later write the string to a log file. I think the problems we are running
into are related to either having allocatable arrays or large numbers of
strings in the namelist, but as a first sanity check I want to print the simple
namelist from this program to a string and then to stdout.